agent-api/requirements.txt
# Core dependencies
langgraph>=0.0.40
langgraph-checkpoint-sqlite>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
greenlet>=2.0.0
# PostgreSQL
asyncpg>=0.29.0
psycopg[binary]>=3.1.0
langgraph-checkpoint-postgres>=2.0.0

# LLM - Volcano Engine (Doubao) via OpenAI-compatible API
openai>=1.30.0

# MCP (Model Context Protocol) - acts as client to external MCP servers
mcp>=1.2.0

# Observability
langsmith>=0.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.1
email-validator>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
alembic>=1.13.0
